requirements.txt
openai==1.98.0
pandas==2.3.1
protobuf==6.31.1
python-dotenv==1.1.1
Requests==2.32.4
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.28.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0